+++ docker/misk-web/README.md
@@ -36,6 +36,7 @@ Includes `misk-web` script and `@misk/` packages. `misk-web` handles running var
 
 | Date       | Docker/ `squareup/ misk-web` | NPM/ `@misk/ common` | NPM/ `@misk/ components` | NPM/ `@misk/ dev` | NPM/ `@misk/ tslint` | Release Notes                                                                                                             |
 | ---------- | ---------------------------- | -------------------- | ------------------------ | ----------------- | -------------------- | ------------------------------------------------------------------------------------------------------------------------- |
+| 2018-11-15 | `0.0.6`                      | `0.0.67`             | `0.0.82`                 | `0.0.68`          | `0.0.10`             | Large changes to Topbar Component, color Enum added with standard colors.                                                 |
 | 2018-11-07 | `0.0.5`                      | `0.0.61`             | `0.0.79`                 | `0.0.64`          | `0.0.10`             | Includes CSS to support all `@blueprintjs` packages.                                                                      |
 | 2018-11-06 | `0.0.4`                      | `0.0.60`             | `0.0.76`                 | `0.0.61`          | `0.0.10`             | Includes all `@blueprintjs` packages.                                                                                     |
 | 2018-11-05 | `0.0.3`                      | `0.0.59`             | `0.0.76`                 | `0.0.60`          | `0.0.10`             | Assumes `node_modules` installed centrally in `web/node_modules`. This will involve updating `tsconfig.json` in each tab. |
