+++ misk/web/@misk/common/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@misk/common",
-  "version": "0.0.65",
+  "version": "0.0.67",
   "license": "SEE LICENSE IN https://github.com/square/misk",
   "description": "Microservice Kontainer Common Libraries, Externals, Styles",
   "author": "Square/Misk Authors (https://github.com/square/misk/graphs/contributors)",
@@ -44,21 +44,21 @@
     "history": "^4.7.2",
     "immutable": "^4.0.0-rc.12",
     "lodash-es": "^4.17.11",
-    "react": "^16.6.0",
-    "react-dom": "^16.6.0",
+    "react": "^16.6.3",
+    "react-dom": "^16.6.3",
     "react-helmet": "^5.2.0",
     "react-hot-loader": "^4.3.12",
-    "react-redux": "^5.1.0",
+    "react-redux": "^5.1.1",
     "react-router": "^4.3.1",
     "react-router-dom": "^4.3.1",
     "react-transition-group": "^2.5.0",
     "redux": "^4.0.1",
     "redux-saga": "^0.16.2",
     "skeleton-css": "^2.0.4",
-    "styled-components": "^4.0.3"
+    "styled-components": "^4.1.1"
   },
   "devDependencies": {
-    "@misk/dev": "^0.0.67",
+    "@misk/dev": "^0.0.68",
     "@misk/tslint": "^0.0.10"
   },
   "miskCachedUrls": {
