+++ misk/web/@misk/components/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@misk/components",
-  "version": "0.0.81",
+  "version": "0.0.83",
   "license": "SEE LICENSE IN https://github.com/square/misk",
   "description": "Microservice Kontainer Common Components",
   "author": "Square/Misk Authors (https://github.com/square/misk/graphs/contributors)",
@@ -31,10 +31,10 @@
     "updateDevCache": "cd ..; ./updateLocalCache.sh"
   },
   "dependencies": {
-    "@misk/common": "^0.0.65"
+    "@misk/common": "^0.0.67"
   },
   "devDependencies": {
-    "@misk/dev": "^0.0.67",
+    "@misk/dev": "^0.0.68",
     "@misk/tslint": "^0.0.10"
   },
   "miskTab": {
