+++ misk/web/@misk/components/src/components/index.ts
@@ -2,4 +2,6 @@ export * from "./ErrorCalloutComponent"
 export * from "./OfflineComponent"
 export * from "./PathDebugComponent"
 export * from "./SidebarComponent"
+export * from "./TextHTMLOrElementComponent"
 export * from "./TopbarComponent"
+export * from "./TopbarComponentDimensionAware"
