+++ misk/web/@misk/components/src/containers/FlexContainer.tsx
@@ -10,4 +10,11 @@ export const FlexContainer = styled.div`
   display: flex;
   flex-direction: row;
   flex-wrap: wrap;
+
+  * {
+    flex-grow: 1;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+  }
 `
