+++ misk/web/@misk/dev/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@misk/dev",
-  "version": "0.0.67",
+  "version": "0.0.68",
   "license": "SEE LICENSE IN https://github.com/square/misk",
   "description": "Microservice Kontainer Development Libraries",
   "author": "Square/Misk Authors (https://github.com/square/misk/graphs/contributors)",
@@ -19,16 +19,16 @@
   },
   "dependencies": {
     "@types/lodash-es": "^4.17.1",
-    "@types/node": "^10.12.5",
-    "@types/react": "^16.7.3",
+    "@types/node": "^10.12.8",
+    "@types/react": "^16.7.6",
     "@types/react-dom": "^16.0.9",
     "@types/react-helmet": "^5.0.7",
     "@types/react-hot-loader": "^4.1.0",
     "@types/react-redux": "^6.0.9",
     "@types/react-router": "^4.4.1",
     "@types/react-router-dom": "^4.3.1",
-    "@types/styled-components": "^4.0.3",
-    "@types/webpack": "^4.4.18",
+    "@types/styled-components": "^4.1.0",
+    "@types/webpack": "^4.4.19",
     "@types/webpack-env": "^1.13.6",
     "awesome-typescript-loader": "^5.2.1",
     "copy-webpack-plugin": "^4.6.0",
