+++ misk/web/tabs/config/package.json
@@ -14,11 +14,11 @@
     "test": "echo no test"
   },
   "dependencies": {
-    "@misk/common": "^0.0.65",
-    "@misk/components": "^0.0.81"
+    "@misk/common": "^0.0.67",
+    "@misk/components": "^0.0.83"
   },
   "devDependencies": {
-    "@misk/dev": "^0.0.67",
+    "@misk/dev": "^0.0.68",
     "@misk/tslint": "^0.0.10"
   },
   "miskTab": {
